AI-LLM/AI应用记忆管理完整指南/requirements.txt
mem0
langchain-openai
langchain-core
langchain-community
python-dotenv
chromadb